# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Quantum Commerce API
  description: API oficial do case fictício Quantum Commerce — disciplina Tools, Automations and Workflows (FIAP). Backend
    determinístico para agentes n8n, Dify, Power Automate e Python. Sem LLM no backend.
  version: 2.0.0
paths:
...
  /api/v1/support/cases:
    post:
      tags:
      - Support
      summary: Create Support Case
      operationId: create_support_case
      parameters:
      - name: X-Lab-Group
        in: header
        required: false
        schema:
          type: string
          default: anonymous
          title: X-Lab-Group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupportCaseCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupportCaseResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/support/cases/{case_id}:
    get:
      tags:
      - Support
      summary: Get Support Case
      operationId: get_support_case
      parameters:
      - name: case_id
        in: path
        required: true
        schema:
          type: string
          title: Case Id
      - name: X-Lab-Group
        in: header
        required: false
        schema:
          type: string
          default: anonymous
          title: X-Lab-Group
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupportCaseDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/customers/{customer_id}/support-cases:
    get:
      tags:
      - Support
      summary: List Customer Support Cases
      operationId: list_customer_support_cases
      parameters:
      - name: customer_id
        in: path
        required: true
        schema:
          type: string
          title: Customer Id
      - name: X-Lab-Group
        in: header
        required: false
        schema:
          type: string
          default: anonymous
          title: X-Lab-Group
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SupportCaseDetail'
                title: Response List Customer Support Cases
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/support/priority/check:
    post:
      tags:
      - Support
      summary: Check Support Priority
      operationId: check_support_priority
      parameters:
      - name: X-Lab-Group
        in: header
        required: false
        schema:
          type: string
          default: anonymous
          title: X-Lab-Group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PriorityCheckRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Check Support Priority
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PriorityCheckRequest:
      properties:
        category:
          type: string
          title: Category
        impact:
          type: string
          title: Impact
        urgency:
          type: string
          title: Urgency
      type: object
      required:
      - category
      - impact
      - urgency
      title: PriorityCheckRequest
...
    SupportCaseCreate:
      properties:
        customer_id:
          type: string
          title: Customer Id
        category:
          type: string
          title: Category
        summary:
          type: string
          title: Summary
        description:
          type: string
          title: Description
      type: object
      required:
      - customer_id
      - category
      - summary
      - description
      title: SupportCaseCreate
    SupportCaseDetail:
      properties:
        case_id:
          type: string
          title: Case Id
        protocol:
          type: string
          title: Protocol
        customer_id:
          type: string
          title: Customer Id
        category:
          type: string
          title: Category
        summary:
          type: string
          title: Summary
        description:
          type: string
          title: Description
        status:
          type: string
          title: Status
        priority:
          type: string
          title: Priority
        lab_group:
          type: string
          title: Lab Group
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Updated At
      type: object
      required:
      - case_id
      - protocol
      - customer_id
      - category
      - summary
      - description
      - status
      - priority
      - lab_group
      - created_at
      title: SupportCaseDetail
    SupportCaseResponse:
      properties:
        case_id:
          type: string
          title: Case Id
        status:
          type: string
          title: Status
        protocol:
          type: string
          title: Protocol
      type: object
      required:
      - case_id
      - status
      - protocol
      title: SupportCaseResponse
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...